--- src/components/error/index.js
@@ -1,3 +1,0 @@
-export { default as GlobalErrorBoundary } from './GlobalErrorBoundary';
-export { default as ApiErrorBoundary } from './ApiErrorBoundary';
-export { default as RenderErrorBoundary } from './RenderErrorBoundary';
