+++ src/errorBoundary/apiErrorBoundary/ApiErrorBoundary.jsx
@@ -1,8 +1,8 @@
 import { CustomButton } from '@/components/customButton';
 import { UI_ERRORS } from '@/constants/errors';
 import { STATUS_CODES } from '@/constants/statusCodes';
+import * as S from '@/errorBoundary/styles';
 import { useRouteError } from 'react-router-dom';
-import * as S from './errorBoundary.styles';
 
 /**
  * ApiErrorBoundary
