+++ src/errorBoundary/index.js
@@ -1,0 +1,3 @@
+export { default as GlobalErrorBoundary } from './globalErrorBoundary';
+export { default as ApiErrorBoundary } from './apiErrorBoundary';
+export { default as RenderErrorBoundary } from './renderErrorBoundary';
