+++ src/pages/error/notFound/index.js
@@ -1,1 +1,1 @@
-export { default as NotFoundPage } from './NotFoundPage';
+export { default } from './NotFoundPage';
